style: :truck: rename old pages in order to enforce consistency
